True 1x Entities
Updated some of the PIP sight settings to get closer to true 1x at the lowest magnification level. Eye relief in the underlying blender model also adjusted to fine tune this, as well as to try and help get rid of the brightness flash. Updated settings help with brightness flash, but it still exists.